Rename `serial` to `Serial`
