+++ Kernel/Devices/Serial/rs232.cpp
@@ -14,7 +14,7 @@
 
 #include <stdarg.h>
 #include <Arch/Arch.hpp>
-#include <Devices/serial/rs232.hpp>
+#include <Devices/Serial/rs232.hpp>
 #include <Memory/heap.hpp>
 #include <Library/stdio.hpp>
 #include <Library/mutex.hpp>
