+++ Kernel/Panic.cpp
@@ -11,7 +11,7 @@
 #include <Devices/Graphics/console.hpp>
 #include <Devices/Graphics/framebuffer.hpp>
 #include <Devices/Graphics/graphics.hpp>
-#include <Devices/serial/rs232.hpp>
+#include <Devices/Serial/rs232.hpp>
 #include <Library/stdio.hpp>
 #include <Panic.hpp>
 #include <Stacktrace.hpp>
