+++ Kernel/Stacktrace.cpp
@@ -11,7 +11,7 @@
 
 #include <Arch/Arch.hpp>
 #include <Devices/Graphics/console.hpp>
-#include <Devices/serial/rs232.hpp>
+#include <Devices/Serial/rs232.hpp>
 #include <Library/stdio.hpp>
 #include <Memory/paging.hpp>
 #include <Stacktrace.hpp>
